+++ R/set_golem_options.R
@@ -46,6 +46,7 @@ set_golem_options <- function(
 	talkative = TRUE,
 	config_file = golem::get_current_config(golem_wd)
 ) {
+	warn_if_in_prod_mode()
 	# TODO here we'll run the
 	# golem_install_dev_pkg() function